Fix and clear

+++ Assets/Editor/NodeEditor/StoryData.cs
@@ -49,19 +49,23 @@ namespace GUIInspector
             if (_storyData.nodesData != null) EditorGUILayout.LabelField("В сюжете учавствует : " + _storyData.nodesData.Count + " Глав");
             else EditorGUILayout.Space();
 
-            for (int i = 0; i < _storyData.nodesData.Count; i++)
+            if (_storyData.nodesData != null)
             {
-                GUI.backgroundColor = Color.white;
-                EditorGUILayout.BeginHorizontal("TextArea");
-                EditorGUILayout.LabelField(_storyData.nodesData[i].name);
-                GUI.backgroundColor = Color.red;
-                if (GUILayout.Button("Удалить", GUILayout.Width(70)))
+                for (int i = 0; i < _storyData.nodesData.Count; i++)
                 {
-                    AssetDatabase.DeleteAsset(AssetDatabase.GetAssetPath(_storyData.nodesData[i]));
-                    _storyData.nodesData.RemoveAt(i);
+                    GUI.backgroundColor = Color.white;
+                    EditorGUILayout.BeginHorizontal("TextArea");
+                    EditorGUILayout.LabelField(_storyData.nodesData[i].name);
+                    GUI.backgroundColor = Color.red;
+                    if (GUILayout.Button("Удалить", GUILayout.Width(70)))
+                    {
+                        AssetDatabase.DeleteAsset(AssetDatabase.GetAssetPath(_storyData.nodesData[i]));
+                        _storyData.nodesData.RemoveAt(i);
+                    }
+                    EditorGUILayout.EndHorizontal();
                 }
-                EditorGUILayout.EndHorizontal();
             }
+            else _storyData.nodesData = new List<GamePart>();
 
             EditorGUILayout.EndVertical();
         }
